requirements.txt
termcolor
numpy
keras
keras-gpu
tensorflow-gpu
coloredlogs